Create README.md
VSC6802 - PHY API
The product pages can be found on the Microchip website at https://www.microchip.com/wwwproducts/en/VSC6802.
The Documents tab contains all the important reference materials for the product.
The current PHY API release for VSC6802 is 4.67.05. It is in maintenance mode and is not expected to change. This
version of the code includes 1588 OOS Recovery for legacy versions of the VSC PHYs.
The VSC6802 PHY API is available on GitHub. Currently, the GitHub repository for PHY API contains the following PHY
API releases: 4.67.03, 4.67.04, and 4.67.05.
The repository is located at https://github.com/microchip-ung/unified_api_4x.
The latest PHY API version is the default version at any given time. For obtaining earlier versions, go to the Tags tab
and choose the associated commit for the earlier version of the code.

The repository is located at https://github.com/microchip-ung/unified_api_4x and you should see a page as shown below:

To Download the latest API release: 4.67.05, Simply Click on the “Download Code” green Button shown on this page.


To Download API release: 4.67.04 or 4.67.03, Follow this procedure:

Step 1: Click on the “5 Commits” that is shown on the right side of the screen: on Apr 23   5 commits
This will display a Page that shows all available releases.

At this point, you need to Select the Version that you wish to download.
Step 2: In this Example, the download of API 4.67.03 is shown (download of API 4.67.04 is shown after this example). 
For API 4.67.03, On the right side, you see the 440a05e .... click on it.   

Step 3: Select the “Browse Files” Button at the Top Right portion of the screen.

Step 4: To Download the API release: 4.67.03, Simply Click on the “Download Code” green Button.

Step 5: The API 4.67.03 has been downloaded to your computer and you are done.



 
If API 4.67.04 was desired, Step 2 would have been:
Step 2: In this Example, I will download API 4.67.04.  On the right side, you see the 03b1abd .... click on it.   

Step 3: Select the “Browse Files” Button at the Top Right portion of the screen.

Step 4: To Download the API release: 4.67.04, Simply Click on the “Download Code” green Button.

Step 5: The API 4.67.04 has been downloaded to your computer and you are done.